requirements.txt
category_encoders==2.5.0
geopy==2.2.0
imbalanced_learn==0.9.1
imblearn==0.0
ipython==8.4.0
joblib==1.1.0
lightgbm==3.3.2
matplotlib==3.5.3
missingno==0.5.1
numpy==1.21.5
pandas==1.4.2
PyYAML==6.0
scikit_learn==1.1.2
seaborn==0.11.2
xgboost==1.6.2
ipykernel==6.15.2 
